Added wikitext downloading and template stripping. Stripped main.py for rework of function.
